bug fix #47387 test table exists before droping table

+++ Database/TeleoptiCCC7/Releases/00000916.sql
@@ -4,8 +4,9 @@
 --Desc: Recreate table ExternalPerformanceData to add Id column
 ----------------  
 
-
-DROP TABLE [dbo].[ExternalPerformanceData]
+IF EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'[dbo].[ExternalPerformanceData]') AND type in (N'U'))
+   DROP TABLE [dbo].[ExternalPerformanceData]
+GO
 
 CREATE TABLE [dbo].[ExternalPerformanceData](
 	[Id] [uniqueidentifier] NOT NULL,
